+++ config/invitations.php
@@ -8,9 +8,11 @@ return [
     'purge' => [
 
         /**
-         * Determines how old invitations should be before purging them.
+         * Default cutoff (in days) for `invitations:purge`: how long past
+         * `expires_at` an invitation must be before being purged.
          *
-         * Set to false if you don't want to remove expired invitations.
+         * Set to false if you don't want purging to remove anything by default.
+         * Override per-run with the `--days=` command argument.
          */
         'expiration_in_days' => 30,
     ],